MZ_Commercial_Logic_v1.0/requirements.txt
# MZ Commercial Area Analysis App Dependencies
PySide6>=6.4.0
requests>=2.28.0
python-dotenv>=1.0.0
pandas>=1.5.0
